Add routed frontend layout with header
