+++ frontend/src/index.css
@@ -1,7 +1,8 @@
 :root {
-  --bg-color: #f5f5f5;
-  --text-color: #213547;
-  --button-bg: #1a1a1a;
+  /* Soft blue-gray palette for a professional look */
+  --bg-color: #eef2f6;
+  --text-color: #1f2937;
+  --button-bg: #004b8d;
   --container-bg: #ffffff;
 
   font-family: system-ui, Avenir, Helvetica, Arial, sans-serif;
@@ -36,6 +37,7 @@ body {
   min-height: 100vh;
   color: var(--text-color);
   background-color: var(--bg-color);
+  font-size: 1.125rem;
 }
 
 h1 {
